Update Server
Updated smtp() function to now prompt the user for email credentials when the client's memory usage statistic (passed over socket) exceeds 50%. If all of this holds true an email 'test' will be mailed to the corresponding client's email address (also configured in the xml). Right now only supports up to 2 clients.

The email credentials required is from the server admin (whoever is executing the server script) and is configured under the <server> tag in the config file (only supports gmail for right now).